Moved VMXHandler
Relocated the header file that handles VMX functions for the wpi to the new Translator sub-directory of Mau
