+++ hal/src/main/native/mau/CAN.cpp
@@ -6,7 +6,7 @@
 /*----------------------------------------------------------------------------*/
 
 #include "HAL/CAN.h"
-#include "VMXHandler.h"
+#include "Translator/VMXHandler.h"
 #include <VMXCAN.h>
 
 namespace hal {
