+++ hal/src/main/native/mau/DIO.cpp
@@ -14,7 +14,7 @@
 #include "HAL/handles/LimitedHandleResource.h"
 #include "HALInitializer.h"
 #include "PortsInternal.h"
-#include "VMXHandler.h"
+#include "Translator/VMXHandler.h"
 #include <VMXIO.h>
 #include <VMXResource.h>
